Update 74HC595_register_and_MAX6675.cpp
change the data type (float => double)

+++ src/74HC595_register_and_MAX6675.cpp
@@ -63,7 +63,7 @@ namespace MAX6675
     digitalWrite(NANO_CLK_PIN, HIGH);
   }
 
-  float readCelsius() 
+  double readCelsius() 
   {
     uint16_t v;
